requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.3
opencv-python==4.5.2.52
pika==1.2.0
Pillow==8.2.0
python-dotenv==0.17.1
requests==2.25.1
torchaudio==0.8.1
typing-extensions==3.10.0.0
urllib3==1.26.4
voluptuous==0.12.1